requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.11
django-redis==5.2.0
django-rq==2.5.1
djangorestframework==3.13.1
importlib-metadata==4.8.3
packaging==21.3
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
rq==1.10.1
sqlparse==0.4.2
typing_extensions==4.0.1
wrapt==1.13.3
zipp==3.6.0
